I WANT TO SLEEP

+++ sh/ghGenerate.sh
@@ -42,7 +42,17 @@ find . -type f | while IFS= read -r file; do
 	if [ -f "$file" ]; then
 		# Is a file
 		if [ "$(echo "$file" | grep -E "$COMPRESS_CRIT")" != "" ]; then
-			rm "$file"
+			if [[ "$file" == *"/index.htm" ]]; then
+				echo "File \"${file}\" is preserved."
+			elif [[ "$file" == *"/index.html" ]]; then
+				echo "File \"${file}\" is preserved."
+			elif [[ "$file" == *"/default.htm" ]]; then
+				echo "File \"${file}\" is preserved."
+			elif [[ "$file" == *"/default.html" ]]; then
+				echo "File \"${file}\" is preserved."
+			else
+				rm "$file"
+			fi
 		else
 			echo "File \"${file}\" is preserved."
 		fi
